[Refactor] Writing.jsx 파일 함수 분리(기능별)/ option 코드 변경 - useTextarea - useUploadImage (2023/09/18)
